chore: move tag page utils to tags utils
