--- openmetadata-ui/src/main/resources/ui/src/pages/tags/TagsPageUtils.tsx
@@ -1,33 +1,0 @@
-/*
- *  Copyright 2022 Collate.
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *  http://www.apache.org/licenses/LICENSE-2.0
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-import { CheckOutlined } from '@ant-design/icons';
-import Loader from 'components/Loader/Loader';
-import React from 'react';
-import { ReactComponent as DeleteIcon } from '../../assets/svg/ic-delete.svg';
-import { DeleteTagsType } from './TagsPage.interface';
-
-export const getDeleteIcon = (
-  deleteTags: DeleteTagsType,
-  id: string | undefined
-) => {
-  if (deleteTags.data?.id === id) {
-    if (deleteTags.data?.status === 'success') {
-      return <CheckOutlined />;
-    }
-
-    return <Loader size="small" type="default" />;
-  }
-
-  return <DeleteIcon name="Delete" width={16} />;
-};
